// app/game/types/game.ts
export interface Lander {
  x: number;
  y: number;
  width: number;
  height: number;
  angle: number;
  velocityX: number;
  velocityY: number;
  rotationSpeed: number;
  fuel: number;
  thrusting: boolean;
  crashed: boolean;
  landed: boolean;
}

export interface TerrainPoint {
  x: number;
  y: number;
}

export interface LandingZone {
  x: number;
  y: number;
  width: number;
}

export interface KeysState {
  [key: string]: boolean;
}

export interface GameLoopOptions {
  enabled: boolean;
  onUpdate: (delta: number) => void;
} 